rename product collection folder name.
